+++ mapsforge-map-android/src/main/java/org/mapsforge/map/android/graphics/AndroidGraphicFactory.java
@@ -4,6 +4,7 @@
  * Copyright 2014-2021 devemux86
  * Copyright 2017 usrusr
  * Copyright 2018 Adrian Batzill
+ * Copyright 2025 Sublimis
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -30,6 +31,7 @@ import org.mapsforge.core.mapelements.PointTextContainer;
 import org.mapsforge.core.mapelements.SymbolContainer;
 import org.mapsforge.core.model.BoundingBox;
 import org.mapsforge.core.model.Point;
+import org.mapsforge.core.util.Parameters;
 import org.mapsforge.map.model.DisplayModel;
 
 import java.io.*;
@@ -49,18 +51,15 @@ public final class AndroidGraphicFactory implements GraphicFactory {
     // if true RESOURCE_BITMAPS will be kept in the cache to avoid
     // multiple loading
     public static final boolean KEEP_RESOURCE_BITMAPS = true;
-    public static final Config NON_TRANSPARENT_BITMAP = Config.RGB_565;
 
-    // determines size of bitmaps used, RGB_565 is 2 bytes per pixel
-    // while ARGB_8888 uses 4 bytes per pixel (with severe impact
-    // on memory use) and allows transparencies. Use ARGB_8888 whenever
-    // you have transparencies in any of the bitmaps. ARGB_4444 is deprecated
-    // and is much slower despite smaller size that ARGB_8888 as it
-    // passes through unoptimized path in the skia library.
+    // see getNonTransparentBitmapConfig
     public static final Config TRANSPARENT_BITMAP = Config.ARGB_8888;
 
-
     public static final Config MONO_ALPHA_BITMAP = Config.ALPHA_8;
+
+    public static final String LIBRARY_FILE_NAME = "mapsforge";
+    private final Object cacheFolderSync = new Object();
+    private volatile File cacheFolder = null;
 
     public static android.graphics.Bitmap convertToAndroidBitmap(Drawable drawable) {
         android.graphics.Bitmap bitmap;
@@ -155,6 +154,18 @@ public final class AndroidGraphicFactory implements GraphicFactory {
         return ((AndroidMatrix) matrix).matrix;
     }
 
+    /**
+     * Determines size of bitmaps used, RGB_565 is 2 bytes per pixel
+     * while ARGB_8888 uses 4 bytes per pixel (with possible impact
+     * on memory use) and allows transparencies. Use ARGB_8888 whenever
+     * you have transparencies in any of the bitmaps. ARGB_4444 is deprecated
+     * and is much slower despite smaller size that ARGB_8888 as it
+     * passes through unoptimized path in the skia library.
+     */
+    public Config getNonTransparentBitmapConfig() {
+        return Parameters.ANDROID_32BIT_COLOR ? Config.ARGB_8888 : Config.RGB_565;
+    }
+
     static android.graphics.Path getPath(Path path) {
         return ((AndroidPath) path).path;
     }
@@ -188,7 +199,7 @@ public final class AndroidGraphicFactory implements GraphicFactory {
         if (isTransparent) {
             return new AndroidBitmap(width, height, TRANSPARENT_BITMAP);
         }
-        return new AndroidBitmap(width, height, NON_TRANSPARENT_BITMAP);
+        return new AndroidBitmap(width, height, getNonTransparentBitmapConfig());
     }
 
     @Override
@@ -268,7 +279,7 @@ public final class AndroidGraphicFactory implements GraphicFactory {
         if (this.svgCacheDir != null) {
             return new File(this.svgCacheDir, name).delete();
         }
-        return this.context.deleteFile(name);
+        return getCacheFile(name).delete();
     }
 
     /*
@@ -278,7 +289,7 @@ public final class AndroidGraphicFactory implements GraphicFactory {
         if (this.svgCacheDir != null) {
             return this.svgCacheDir.list();
         }
-        return this.context.fileList();
+        return getCacheFile("").list();
     }
 
     /*
@@ -288,7 +299,7 @@ public final class AndroidGraphicFactory implements GraphicFactory {
         if (this.svgCacheDir != null) {
             return new FileInputStream(new File(this.svgCacheDir, name));
         }
-        return this.context.openFileInput(name);
+        return new FileInputStream(getCacheFile(name));
     }
 
     /*
@@ -298,7 +309,7 @@ public final class AndroidGraphicFactory implements GraphicFactory {
         if (this.svgCacheDir != null) {
             return new FileOutputStream(new File(this.svgCacheDir, name), mode == Context.MODE_APPEND);
         }
-        return this.context.openFileOutput(name, mode);
+        return new FileOutputStream(getCacheFile(name), mode == Context.MODE_APPEND);
     }
 
     @Override
@@ -319,5 +330,21 @@ public final class AndroidGraphicFactory implements GraphicFactory {
 
     public void setSvgCacheDir(File svgCacheDir) {
         this.svgCacheDir = svgCacheDir;
+    }
+
+    public File getCacheFile(String filename) {
+        File ourCacheFolder = this.cacheFolder;
+        if (ourCacheFolder == null) {
+            synchronized (this.cacheFolderSync) {
+                ourCacheFolder = this.cacheFolder;
+                if (ourCacheFolder == null) {
+                    ourCacheFolder = new File(this.context.getCacheDir(), LIBRARY_FILE_NAME);
+                    ourCacheFolder.mkdirs();
+                    this.cacheFolder = ourCacheFolder;
+                }
+            }
+        }
+
+        return new File(ourCacheFolder, filename);
     }
 }